+++ Bee.Base/ImageConsts.cs
@@ -1,0 +1,17 @@
+
+namespace Bee.Base;
+
+/// <summary>
+/// 资源常量
+/// </summary>
+public class AssetConsts
+{
+    /// <summary>
+    /// 文件夹图片
+    /// </summary>
+    public static Uri Folder => new("avares://Bee/Assets/images/folder-80.png");
+    /// <summary>
+    /// Uml图片
+    /// </summary>
+    public static Uri Uml => new("avares://Bee/Assets/images/uml-80.png");
+}