+++ Bee.Base/Models/Tasks/TaskExtensions.cs
@@ -6,19 +6,19 @@ public static class TaskExtensions
     /// <summary>
     /// 合并图像集合
     /// </summary>
-    /// <param name="source">源图像对象集合</param>
-    /// <param name="newImages">要合并至源集合对象的新集合</param>
-    public static void Merge(this ICollection<TaskItem> source, ICollection<TaskItem> newImages)
+    /// <param name="source">源对象集合</param>
+    /// <param name="newItems">要合并至源集合对象的新集合</param>
+    public static void Merge(this ICollection<TaskItem> source, ICollection<TaskItem> newItems)
     {
-        if (newImages == null)
+        if (newItems == null)
         {
             return;
         }
 
-        var existingSources = source.Select(i => i.FileName).ToHashSet();
-        foreach (var newItem in newImages)
+        var existingSources = source.Select(i => i.Input).ToHashSet();
+        foreach (var newItem in newItems)
         {
-            if (!existingSources.Contains(newItem.FileName))
+            if (!existingSources.Contains(newItem.Input))
             {
                 source.Add(newItem);
             }